Rely on Rx for async completion flow
